requirements.txt
numpy<2
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-surprise>=1.1.3